requirements.txt
pandas == 2.0.3
numpy == 1.25.0
pymysql == 1.1.0
mysql_connector ==  9.2.0